loader: more avoidance of sized types where not needed

+++ src/loader.cc
@@ -45,7 +45,7 @@ assert_rdt_abi()
 bool
 save_tree(RDTree* tree, const char* filename)
 {
-  uint32_t n_nodes;
+  int n_nodes;
   bool success;
   FILE* output;
 
@@ -62,15 +62,15 @@ save_tree(RDTree* tree, const char* filename)
       goto save_tree_close;
     }
 
-  n_nodes = (uint32_t)roundf(powf(2.f, tree->header.depth)) - 1;
-  if (fwrite(tree->nodes, sizeof(Node), n_nodes, output) != n_nodes)
+  n_nodes = roundf(powf(2.f, tree->header.depth)) - 1;
+  if (fwrite(tree->nodes, sizeof(Node), n_nodes, output) != (size_t)n_nodes)
     {
       fprintf(stderr, "Error writing tree nodes\n");
       goto save_tree_close;
     }
 
   if (fwrite(tree->label_pr_tables, sizeof(float) * tree->header.n_labels,
-             tree->n_pr_tables, output) != tree->n_pr_tables)
+             tree->n_pr_tables, output) != (size_t)tree->n_pr_tables)
     {
       fprintf(stderr, "Error writing tree probability tables\n");
       goto save_tree_close;
@@ -196,8 +196,8 @@ count_pr_tables(JSON_Object* node)
 }
 
 static void
-unpack_json_tree(JSON_Object* jnode, Node* nodes, uint32_t node_index,
-                 float* pr_tables, uint32_t* table_index, uint8_t n_labels)
+unpack_json_tree(JSON_Object* jnode, Node* nodes, int node_index,
+                 float* pr_tables, int* table_index, int n_labels)
 {
   Node* node = &nodes[node_index];
 
@@ -206,7 +206,7 @@ unpack_json_tree(JSON_Object* jnode, Node* nodes, uint32_t node_index,
       float* pr_table = &pr_tables[(*table_index) * n_labels];
 
       JSON_Array* p = json_object_get_array(jnode, "p");
-      for (uint8_t i = 0; i < n_labels; i++)
+      for (int i = 0; i < n_labels; i++)
         {
           pr_table[i] = (float)json_array_get_number(p, i);
         }
@@ -219,11 +219,11 @@ unpack_json_tree(JSON_Object* jnode, Node* nodes, uint32_t node_index,
   JSON_Array* u = json_object_get_array(jnode, "u");
   JSON_Array* v = json_object_get_array(jnode, "v");
 
-  node->uv[0] = (float)json_array_get_number(u, 0);
-  node->uv[1] = (float)json_array_get_number(u, 1);
-  node->uv[2] = (float)json_array_get_number(v, 0);
-  node->uv[3] = (float)json_array_get_number(v, 1);
-  node->t = (float)json_object_get_number(jnode, "t");
+  node->uv[0] = json_array_get_number(u, 0);
+  node->uv[1] = json_array_get_number(u, 1);
+  node->uv[2] = json_array_get_number(v, 0);
+  node->uv[3] = json_array_get_number(v, 1);
+  node->t = json_object_get_number(jnode, "t");
   node->label_pr_idx = 0;
 
   unpack_json_tree(json_object_get_object(jnode, "l"), nodes,
@@ -233,7 +233,7 @@ unpack_json_tree(JSON_Object* jnode, Node* nodes, uint32_t node_index,
 }
 
 RDTree*
-load_json_tree(uint8_t* json_tree_buf, uint32_t len)
+load_json_tree(uint8_t* json_tree_buf, int len)
 {
   assert_rdt_abi();
 
@@ -283,13 +283,13 @@ load_json_tree(uint8_t* json_tree_buf, uint32_t len)
   tree->n_pr_tables = n_pr_tables;
 
   // Allocate tree structure
-  uint32_t n_nodes = (uint32_t)roundf(powf(2.f, tree->header.depth)) - 1;
+  int n_nodes = roundf(powf(2.f, tree->header.depth)) - 1;
   tree->nodes = (Node*)xmalloc(n_nodes * sizeof(Node));
   tree->label_pr_tables = (float*)
     xmalloc(n_pr_tables * tree->header.n_labels * sizeof(float));
 
   // Copy over nodes and probability tables
-  uint32_t table_index = 0;
+  int table_index = 0;
   unpack_json_tree(root, tree->nodes, 0, tree->label_pr_tables, &table_index,
                    tree->header.n_labels);
 
@@ -315,13 +315,13 @@ read_json_tree(const char* filename)
 }
 
 RDTree*
-load_tree(uint8_t* tree_buf, uint32_t len)
+load_tree(uint8_t* tree_buf, int len)
 {
   assert_rdt_abi();
 
   RDTree* tree = (RDTree*)xcalloc(1, sizeof(RDTree));
 
-  if (len < sizeof(RDTHeader))
+  if ((size_t)len < sizeof(RDTHeader))
     {
       fprintf(stderr, "Buffer too small to contain tree\n");
       free_tree(tree);
@@ -341,15 +341,15 @@ load_tree(uint8_t* tree_buf, uint32_t len)
   if (tree->header.version != RDT_VERSION)
     {
       fprintf(stderr, "Incompatible RDT version, expected %u, found %u\n",
-              RDT_VERSION, (uint32_t)tree->header.version);
+              RDT_VERSION, (unsigned)tree->header.version);
       free_tree(tree);
       return NULL;
     }
 
   // Read in the decision tree nodes
-  uint32_t n_nodes = (uint32_t)roundf(powf(2.f, tree->header.depth)) - 1;
+  int n_nodes = roundf(powf(2.f, tree->header.depth)) - 1;
   tree->nodes = (Node*)xmalloc(n_nodes * sizeof(Node));
-  if (len < (sizeof(Node) * n_nodes))
+  if ((size_t)len < (sizeof(Node) * n_nodes))
     {
       fprintf(stderr, "Error parsing tree nodes\n");
       free_tree(tree);
@@ -360,21 +360,21 @@ load_tree(uint8_t* tree_buf, uint32_t len)
   len -= sizeof(Node) * n_nodes;
 
   // Read in the label probabilities
-  long label_bytes = len;
+  int label_bytes = len;
   if (label_bytes % sizeof(float) != 0)
     {
       fprintf(stderr, "Unexpected size of label probability tables\n");
       free_tree(tree);
       return NULL;
     }
-  uint32_t n_prs = label_bytes / sizeof(float);
+  int n_prs = label_bytes / sizeof(float);
   if (n_prs % tree->header.n_labels != 0)
     {
       fprintf(stderr, "Unexpected number of label probabilities\n");
       free_tree(tree);
       return NULL;
     }
-  uint32_t n_tables = n_prs / tree->header.n_labels;
+  int n_tables = n_prs / tree->header.n_labels;
 
   tree->n_pr_tables = n_tables;
   tree->label_pr_tables = (float*)xmalloc(label_bytes);
@@ -414,14 +414,14 @@ free_tree(RDTree* tree)
 }
 
 static RDTree**
-load_any_forest(uint8_t** tree_bufs, uint32_t* tree_buf_lengths,
-                uint32_t n_trees, bool is_json)
+load_any_forest(uint8_t** tree_bufs, int* tree_buf_lengths,
+                int n_trees, bool is_json)
 {
   bool error = false;
-  uint8_t n_labels = 0;
+  int n_labels = 0;
   RDTree** trees = (RDTree**)xcalloc(n_trees, sizeof(RDTree*));
 
-  for (uint32_t i = 0; i < n_trees; i++)
+  for (int i = 0; i < n_trees; i++)
     {
       // Validate the decision tree
       RDTree* tree = trees[i] = is_json ?
@@ -440,9 +440,9 @@ load_any_forest(uint8_t** tree_bufs, uint32_t* tree_buf_lengths,
         }
       if (tree->header.n_labels != n_labels)
         {
-          fprintf(stderr, "Tree %u has %u labels, expected %u\n",
-                  i, (uint32_t)tree->header.n_labels,
-                  (uint32_t)n_labels);
+          fprintf(stderr, "Tree %d has %d labels, expected %d\n",
+                  i, tree->header.n_labels,
+                  n_labels);
           error = true;
           break;
         }
@@ -450,7 +450,7 @@ load_any_forest(uint8_t** tree_bufs, uint32_t* tree_buf_lengths,
 
   if (error)
     {
-      for (uint32_t i = 0; i < n_trees; i++)
+      for (int i = 0; i < n_trees; i++)
         {
           if (trees[i])
             {
@@ -466,27 +466,26 @@ load_any_forest(uint8_t** tree_bufs, uint32_t* tree_buf_lengths,
 }
 
 RDTree**
-load_json_forest(uint8_t** tree_bufs, uint32_t* tree_buf_lengths,
-                 uint32_t n_trees)
+load_json_forest(uint8_t** tree_bufs, int* tree_buf_lengths, int n_trees)
 {
   return load_any_forest(tree_bufs, tree_buf_lengths, n_trees, true);
 }
 
 RDTree**
-load_forest(uint8_t** tree_bufs, uint32_t* tree_buf_lengths, uint32_t n_trees)
+load_forest(uint8_t** tree_bufs, int* tree_buf_lengths, int n_trees)
 {
   return load_any_forest(tree_bufs, tree_buf_lengths, n_trees, false);
 }
 
 static RDTree**
-read_any_forest(const char** files, uint32_t n_files, bool is_json)
+read_any_forest(const char** files, int n_files, bool is_json)
 {
   uint8_t* tree_bufs[n_files];
-  uint32_t tree_buf_lengths[n_files];
-  uint32_t n_trees = 0;
+  int tree_buf_lengths[n_files];
+  int n_trees = 0;
   bool error = false;
 
-  for (uint32_t i = 0; i < n_files; i++)
+  for (int i = 0; i < n_files; i++)
     {
       const char* tree_file = files[i];
 
@@ -525,7 +524,7 @@ read_any_forest(const char** files, uint32_t n_files, bool is_json)
   RDTree** forest = error ?
     NULL : load_any_forest(tree_bufs, tree_buf_lengths, n_trees, is_json);
 
-  for (uint32_t i = 0; i < n_trees; i++)
+  for (int i = 0; i < n_trees; i++)
     {
       xfree(tree_bufs[i]);
     }
@@ -534,13 +533,13 @@ read_any_forest(const char** files, uint32_t n_files, bool is_json)
 }
 
 RDTree**
-read_json_forest(const char** files, uint32_t n_files)
+read_json_forest(const char** files, int n_files)
 {
   return read_any_forest(files, n_files, true);
 }
 
 RDTree**
-read_forest(const char** files, uint32_t n_files)
+read_forest(const char** files, int n_files)
 {
   return read_any_forest(files, n_files, false);
 }
